refactor: Update UI
Because the whole authentication process for AniList will happen on the website
it's not necessary to have text fields for the user's email and passwords. Thus,
text fields, their custom class and all the code that handled keyboard events
were removed. Instead there is just one button now that opens the authentication
URL on AniList in a web view.