fix: Add isAccessTokenValid and validateAccessToken methods
When the access token expired while using the application and a request was made,
the app always shows errors because requests to the AniList API don't work without a
valid access token, thus the app had to be restarted in order to work again.

Those methods should be used in order to check if a valid access token is available on
each request and if not, request a new access token with the refresh token.