feat: Add a tap gesture recognizer that hides the keyboard when the main view is tapped
The keyboard now hides when the main view is tapped. However to ensure that collection view cells
still register taps and thus open a series' details the main view's tap gesture recognizer shouldn't
receive touches that happen inside of one of the collection view's cells because then the main view's
tap gesture recognizer would receive the touch events. Because of this the
gestureRecognizer(gestureRecognizer:shouldReceive:) method was implemented appropriately
in the search view controller which serves as the gesture recognizer's delegate