fix: Modify series detail table view data source so that only cells with available content are shown
Before, all cells were always shown but when data wasn't available an empty table view cell was returned.
This lead to the problem that whitespace was shown when no data was available for a cell which didn't look
very nice.

Now every time the series property gets set, the availableCellTypes array gets filled with all cell
types that have data available for them. The table view's data source then sets the number of cells depending
on the number of items in the availableCellTypes property and in the cellForRowAt:indexPath method an
appropriate cell is returned based on the cell type